% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{argonAvatars}
\alias{argonAvatars}
\title{Create a Boostrap 4 avatar container}
\usage{
argonAvatars(...)
}
\arguments{
\item{...}{Slot for \link{argonAvatar}.}
}
\description{
Build an argon avatar container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
