% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R
\name{argonDashBody}
\alias{argonDashBody}
\title{Create a Boostrap 4 dashboard body}
\usage{
argonDashBody(...)
}
\arguments{
\item{...}{Body content, slot for \link{argonTabItems}}
}
\description{
Build an argon dashboard body
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
