% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{argonDashFooter}
\alias{argonDashFooter}
\title{Create a Boostrap 4 dashboard footer}
\usage{
argonDashFooter(..., copyrights = NULL, src = NULL)
}
\arguments{
\item{...}{Slot for \link{argonFooterMenu}.}

\item{copyrights}{Left text, if any.}

\item{src}{Left text src, if any.}
}
\description{
Build an argon dashboard footer
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
