% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argonDashGallery.R
\name{argonDashGallery}
\alias{argonDashGallery}
\title{Launch the argonDash Gallery}
\usage{
argonDashGallery()
}
\description{
A gallery of all components available in argonDash.
}
\examples{

if (interactive()) {

 argonDashGallery()

}
}
