% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{argonDashNavbar}
\alias{argonDashNavbar}
\title{Create a Boostrap 4 dashboard navbar}
\usage{
argonDashNavbar(...)
}
\arguments{
\item{...}{Any element such as \link{argonDropNav}.}
}
\description{
Build an argon dashboard navbar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
