% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{argonDashSidebar}
\alias{argonDashSidebar}
\title{Create a Boostrap 4 dashboard sidebar}
\usage{
argonDashSidebar(
  ...,
  dropdownMenus = NULL,
  id,
  brand_url = NULL,
  brand_logo = NULL,
  side = c("start", "end"),
  skin = c("default", "white"),
  color = c("primary", "dark", "info", "success", "warning", "danger")
)
}
\arguments{
\item{...}{Any element such as \link{argonDashSidebar}, \link{argonSidebarMenu},
\link{argonSidebarItem}, \link{argonSidebarHeader} and \link{argonSidebarDivider}.}

\item{dropdownMenus}{Slot for dropdown menus. Not displayd on large screens.}

\item{id}{Unique sidebar id.}

\item{brand_url}{Sidebar brand url}

\item{brand_logo}{Sidebar brand logo}

\item{skin}{Sidebar skin. "light" by default.}

\item{color}{Links color.}

\item{size}{Sidebar size: "s", "md", "lg". "md" by default.}
}
\description{
Build an argon dashboard sidebar
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
