% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{argonDropNav}
\alias{argonDropNav}
\title{Create a Boostrap 4 dashboard dropdown menu}
\usage{
argonDropNav(..., title = NULL, src = NULL, orientation = "right")
}
\arguments{
\item{...}{slot for \link{argonDropNavItem}, \link{argonDropNavTitle} or \link{argonDropNavDivider}.}

\item{title}{Menu title.}

\item{src}{Menu image, if any.}

\item{orientation}{Dropdown orientation: "left" or "right".}
}
\description{
Build an argon dashboard dropdown menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
