% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{argonDropNavDivider}
\alias{argonDropNavDivider}
\title{Create a Boostrap 4 dashboard dropdown menu divider}
\usage{
argonDropNavDivider()
}
\description{
Build an argon dashboard dropdown menu divider
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
