% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{argonDropNavItem}
\alias{argonDropNavItem}
\title{Create a Boostrap 4 dashboard dropdown menu item}
\usage{
argonDropNavItem(title = NULL, src = NULL, icon = NULL)
}
\arguments{
\item{title}{Item title.}

\item{src}{External link.}

\item{icon}{Icon, if any. Expect \link{argonIcon} or \link[shiny]{icon}.}
}
\description{
Build an argon dashboard dropdown menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
