% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{argonDropNavTitle}
\alias{argonDropNavTitle}
\title{Create a Boostrap 4 dashboard dropdown menu title}
\usage{
argonDropNavTitle(title = NULL)
}
\arguments{
\item{title}{Menu title.}
}
\description{
Build an argon dashboard dropdown menu title
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
