% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{argonFooterMenu}
\alias{argonFooterMenu}
\title{Create a Boostrap 4 dashboard footer menu}
\usage{
argonFooterMenu(...)
}
\arguments{
\item{...}{Slot for \link{argonFooterItem}.}
}
\description{
Build an argon dashboard footer menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
