% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{argonInfoCard}
\alias{argonInfoCard}
\title{Create a Boostrap 4 dashboard info card}
\usage{
argonInfoCard(
  value,
  title = NULL,
  stat = NULL,
  stat_icon = NULL,
  description = NULL,
  icon,
  icon_background = "default",
  hover_lift = FALSE,
  shadow = FALSE,
  background_color = NULL,
  gradient = FALSE,
  width = 3
)
}
\arguments{
\item{value}{info card value.}

\item{title}{info card title.}

\item{stat}{info card stat value.}

\item{stat_icon}{info card stat icon. Expect \link{argonIcon} or \link[shiny]{icon}.}

\item{description}{info card stat description.}

\item{icon}{info card icon. Expect \link{argonIcon} or \link[shiny]{icon}.}

\item{icon_background}{info card icon background color.}

\item{hover_lift}{Whether to apply a lift effect on hover. FALSE by default.
Not compatible with floating.}

\item{shadow}{Whether to apply a shadow effect. FALSE by default.}

\item{background_color}{Card background color. NULL by default.}

\item{gradient}{Whether to apply a gradient effect on the card background. FALSE by default.}

\item{width}{info card width. Between 1 and 12.}
}
\description{
Build an argon dashboard info card
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
