% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{argonSidebarDivider}
\alias{argonSidebarDivider}
\title{A divider to put inside \link{argonDashSidebar}}
\usage{
argonSidebarDivider()
}
\description{
A divider to put inside \link{argonDashSidebar}
}
