% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{argonSidebarHeader}
\alias{argonSidebarHeader}
\title{A header to put inside \link{argonDashSidebar}}
\usage{
argonSidebarHeader(title = NULL)
}
\arguments{
\item{title}{Header title}
}
\description{
A header to put inside \link{argonDashSidebar}
}
