% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{argonSidebarItem}
\alias{argonSidebarItem}
\title{Create a Boostrap 4 argon sidebar menu}
\usage{
argonSidebarItem(..., tabName = NULL, icon = NULL)
}
\arguments{
\item{...}{Item name.}

\item{tabName}{Should correspond exactly to the tabName given in \code{\link{argonTabItem}}.}

\item{icon}{Item icon. \link{argonIcon} or \link[shiny]{icon}.}
}
\description{
Build an argon dashboard main sidebar menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
