% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{argonSidebarMenu}
\alias{argonSidebarMenu}
\title{Create a Boostrap 4 argon sidebar menu}
\usage{
argonSidebarMenu(...)
}
\arguments{
\item{...}{Slot for \link{argonSidebarItem}.}
}
\description{
Build an argon dashboard main sidebar menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
