% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{argonSidebarMenuItem}
\alias{argonSidebarMenuItem}
\title{Create a Boostrap 4 argon sidebar menu}
\usage{
argonSidebarMenuItem(..., tabName = NULL, icon = NULL,
  icon_color = NULL)
}
\arguments{
\item{...}{Item name.}

\item{tabName}{Should correspond exactly to the tabName given in \code{\link{argonTabItem}}.}

\item{icon}{Item icon.}

\item{icon_color}{Icon color.}
}
\description{
Build an argon dashboard main sidebar menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
