% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R
\name{argonTabItems}
\alias{argonTabItems}
\title{A container for tab items}
\usage{
argonTabItems(...)
}
\arguments{
\item{...}{Items to put in the container. Each item should be a
\code{\link{argonTabItem}}.}
}
\description{
A container for tab items
}
