% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{argonTableItems}
\alias{argonTableItems}
\title{Create a Boostrap 4 table item row}
\usage{
argonTableItems(...)
}
\arguments{
\item{...}{Slot for \link{argonTableItem}.}
}
\description{
Build an argon table item row
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
