% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_app_link}
\alias{create_app_link}
\title{Create shinylive iframe}
\usage{
create_app_link(url, mode = c("app", "editor"), header = TRUE)
}
\arguments{
\item{url}{app url. A shinylive link.}

\item{mode}{How to display the shinylive app. Default to app mode.}

\item{header}{Whether to display the shinylive header. Default to TRUE.}
}
\description{
Useful for pkgdown website
}
\keyword{internal}
