% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{argonDashHeader}
\alias{argonDashHeader}
\title{Create a Boostrap 5 dashboard header}
\usage{
argonDashHeader(height = 300, color = "dark")
}
\arguments{
\item{height}{Header height.}

\item{color}{Background color.}
}
\description{
Build an argon dashboard header
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
