% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{argonDashHeader}
\alias{argonDashHeader}
\title{Create a Boostrap 4 dashboard header}
\usage{
argonDashHeader(..., gradient = TRUE, color = NULL)
}
\arguments{
\item{...}{Any element}

\item{gradient}{Whether to apply a gradient effect. TRUE by default.}

\item{color}{Header color. See \url{https://demos.creative-tim.com/argon-design-system/index.html}.}
}
\description{
Build an argon dashboard header
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
